requirements.txt
anaconda==0.0.1.1
attrs==24.2.0
bidict==0.23.1
breeze_connect==1.0.56
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
idna==3.7
logging==0.4.9.6
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.0.1
outcome==1.3.0.post0
pandas==2.2.2
Pygments==2.18.0
pyotp==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
requests==2.32.3
rich==13.7.1
schedule==1.2.2
selenium==4.23.1
setuptools==72.1.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
ta==0.11.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
